projects/llm-uni-bias/requirements.txt
anthropic>=0.25.0
openai>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.0
python-dotenv>=0.19.0